user-service/requirements.txt
grpcio
grpcio-tools
sqlalchemy>=2.0
asyncpg
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-asyncpg
